+++ feature/analysis/src/commonMain/kotlin/com/akole/dividox/feature/analysis/SecurityDetailViewModel.kt
@@ -29,8 +29,6 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.toLocalDateTime
 import kotlin.time.Clock
 
 class SecurityDetailViewModel(
@@ -138,34 +136,7 @@ class SecurityDetailViewModel(
                     }
                     val events = eventsDeferred.await() ?: return@coroutineScope
                     val pricePoints = pricePointsDeferred.await()
-
-                    val dividendsByYear = events
-                        .groupBy { it.exDividendDate.year }
-                        .mapValues { (_, evts) -> evts.sumOf { it.amountPerShare } }
-
-                    val priceByYear = pricePoints
-                        .groupBy { it.timestamp.toLocalDateTime(TimeZone.UTC).year }
-                        .mapValues { (_, pts) ->
-                            pts.maxByOrNull { it.timestamp }?.close ?: 0.0
-                        }
-
-                    val currentYear = Clock.System.now().toLocalDateTime(TimeZone.UTC).year
-                    val bars = (0..9).mapNotNull { yearOffset ->
-                        val year = currentYear - yearOffset
-                        val payout = dividendsByYear[year] ?: return@mapNotNull null
-                        if (payout <= 0.0) return@mapNotNull null
-                        val price = priceByYear[year]
-                            ?: priceByYear.entries.filter { it.key <= year }
-                                .maxByOrNull { it.key }?.value
-                            ?: return@mapNotNull null
-                        if (price <= 0.0) return@mapNotNull null
-                        DividendGrowthBar(
-                            year = year,
-                            absoluteValue = payout,
-                            percentageOfPrice = (payout / price) * 100,
-                        )
-                    }.reversed()
-
+                    val bars = calculateDividendGrowthBars(events, pricePoints)
                     updateViewState { copy(dividendGrowthData = bars) }
                 }
             }
