+++ feature/dashboard/src/commonMain/kotlin/com/akole/dividox/feature/dashboard/DashboardViewModel.kt
@@ -10,7 +10,6 @@ import com.akole.dividox.common.network.connectivity.NetworkConnectivityManager
 import com.akole.dividox.common.settings.AppRefreshTracker
 import com.akole.dividox.common.settings.domain.usecase.ObserveAppSettingsUseCase
 import com.akole.dividox.common.settings.domain.usecase.SetCurrencyUseCase
-import com.akole.dividox.component.market.domain.model.ChartPeriod as MarketChartPeriod
 import com.akole.dividox.component.watchlist.domain.usecase.RemoveFromWatchlistUseCase
 import com.akole.dividox.feature.dashboard.DashboardContract.DashboardSideEffect
 import com.akole.dividox.feature.dashboard.DashboardContract.DashboardViewEvent
@@ -24,8 +23,6 @@ import com.akole.dividox.feature.dashboard.DashboardContract.DashboardViewState
 import com.akole.dividox.integration.dividend.domain.usecase.GetPeriodDividendsUseCase
 import com.akole.dividox.integration.dividend.domain.usecase.ObservePortfolioChangesUseCase
 import com.akole.dividox.integration.dividend.domain.usecase.SyncDividendHistoryFromHoldingsUseCase
-import com.akole.dividox.integration.security.domain.model.EnrichedWatchlistEntry
-import com.akole.dividox.integration.security.domain.model.PortfolioSummary
 import com.akole.dividox.integration.security.domain.usecase.GetEnrichedWatchlistUseCase
 import com.akole.dividox.integration.security.domain.usecase.GetPortfolioPeriodGainUseCase
 import com.akole.dividox.integration.security.domain.usecase.GetPortfolioSummaryUseCase
@@ -39,11 +36,6 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.launch
-import kotlinx.datetime.DatePeriod
-import kotlinx.datetime.LocalDate
-import kotlinx.datetime.TimeZone
-import kotlinx.datetime.minus
-import kotlinx.datetime.todayIn
 
 @Suppress("LongParameterList", "TooManyFunctions")
 class DashboardViewModel(
@@ -147,8 +139,8 @@ class DashboardViewModel(
                 periodGainFlow,
             ) { summary, watchlist, settings, (dividends, lifetime), (gainAbs, gainPct) ->
                 val targetCurrency = settings.currency
-                val convertedSummary = convertSummary(summary, targetCurrency)
-                val convertedWatchlistPrices = convertWatchlistPrices(watchlist, targetCurrency)
+                val convertedSummary = currencyConverter.convertSummary(summary, targetCurrency)
+                val convertedWatchlistPrices = currencyConverter.convertWatchlistPrices(watchlist, targetCurrency)
                 viewState.value.copy(
                     isLoading = false,
                     isRefreshing = false,
@@ -159,69 +151,15 @@ class DashboardViewModel(
                     convertedSummary = convertedSummary,
                     convertedWatchlistPrices = convertedWatchlistPrices,
                     periodGainPercent = gainPct,
-                    periodGainAbsolute = convertAmount(gainAbs, targetCurrency),
-                    periodDividends = convertAmount(dividends, targetCurrency),
-                    lifetimeDividends = convertAmount(lifetime, targetCurrency),
+                    periodGainAbsolute = currencyConverter.convertAmount(gainAbs, targetCurrency),
+                    periodDividends = currencyConverter.convertAmount(dividends, targetCurrency),
+                    lifetimeDividends = currencyConverter.convertAmount(lifetime, targetCurrency),
                 )
             }.collect { newState ->
                 val now = Clock.System.now()
                 refreshTracker.notifyRefreshed(now)
                 updateViewState { newState.copy(lastUpdated = now) }
             }
-        }
-    }
-
-    private fun ChartPeriod.toStartDate(): LocalDate? {
-        val today = Clock.System.todayIn(TimeZone.UTC)
-        return when (this) {
-            ChartPeriod.ALL -> null
-            ChartPeriod.ONE_DAY -> today.minus(DatePeriod(days = 1))
-            ChartPeriod.ONE_WEEK -> today.minus(DatePeriod(days = 7))
-            ChartPeriod.ONE_MONTH -> today.minus(DatePeriod(months = 1))
-            ChartPeriod.ONE_YEAR -> today.minus(DatePeriod(years = 1))
-            ChartPeriod.YEAR_TO_DATE -> LocalDate(today.year, 1, 1)
-        }
-    }
-
-    private suspend fun convertAmount(amount: Double, to: Currency): Double {
-        if (to == Currency.USD) return amount
-        return currencyConverter.getRate(Currency.USD, to).getOrNull()?.let { rate ->
-            amount * rate
-        } ?: amount
-    }
-
-    private fun ChartPeriod.toMarketPeriod(): MarketChartPeriod = when (this) {
-        ChartPeriod.ONE_DAY -> MarketChartPeriod.ONE_DAY
-        ChartPeriod.ONE_WEEK -> MarketChartPeriod.ONE_WEEK
-        ChartPeriod.ONE_MONTH -> MarketChartPeriod.ONE_MONTH
-        ChartPeriod.ONE_YEAR -> MarketChartPeriod.ONE_YEAR
-        ChartPeriod.YEAR_TO_DATE -> MarketChartPeriod.YTD
-        ChartPeriod.ALL -> MarketChartPeriod.ALL
-    }
-
-    private suspend fun convertSummary(summary: PortfolioSummary, to: Currency): PortfolioSummary {
-        val from = Currency.USD
-        if (from == to) return summary
-        val rate = currencyConverter.getRate(from, to).getOrNull() ?: return summary
-        return PortfolioSummary(
-            totalValue = summary.totalValue * rate,
-            totalGain = summary.totalGain * rate,
-            totalGainPercent = summary.totalGainPercent,
-            totalYield = summary.totalYield,
-            dividendsCollected = summary.dividendsCollected * rate,
-        )
-    }
-
-    private suspend fun convertWatchlistPrices(
-        watchlist: List<EnrichedWatchlistEntry>,
-        to: Currency,
-    ): Map<String, Double> = buildMap {
-        watchlist.forEach { entry ->
-            val quote = entry.quote ?: return@forEach
-            val ticker = entry.entry.tickerId
-            val from = Currency.entries.firstOrNull { it.code == quote.currency } ?: Currency.USD
-            val converted = currencyConverter.convert(quote.price, from, to).getOrElse { quote.price }
-            put(ticker, converted)
         }
     }
